plugins/_gmail_plugin/requirements.txt
beautifulsoup4==4.12.2
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0